Move to JBoss env
