+++ containers/resin4-embedded/src/main/resources/META-INF/services/org.jboss.arquillian.container.resin.embedded_4.ShrinkWrapWebAppContext
