+++ containers/resin4-embedded/src/main/resources/META-INF/services/org.jboss.arquillian.spi.ContainerConfiguration
@@ -1,1 +1,1 @@
-com.caucho.arquillian.container.resin.embedded_4.ResinEmbeddedConfiguration
+org.jboss.arquillian.container.resin.embedded_4.ResinEmbeddedConfiguration
