+++ containers/resin4-embedded/src/main/resources/META-INF/services/org.jboss.arquillian.spi.DeployableContainer
@@ -1,1 +1,1 @@
-com.caucho.arquillian.container.resin.embedded_4.ResinEmbeddedContainer
+org.jboss.arquillian.container.resin.embedded_4.ResinEmbeddedContainer
