+++ containers/resin4-embedded/src/test/java/org/jboss/arquillian/container/resin/embedded_4/MavenArtifactResolver.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.caucho.arquillian.container.resin.embedded_4;
+package org.jboss.arquillian.container.resin.embedded_4;
 
 import java.io.File;
