+++ containers/resin4-embedded/src/test/java/org/jboss/arquillian/container/resin/embedded_4/TestBean.java
@@ -14,31 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.caucho.arquillian.container.resin.embedded_4;
+package org.jboss.arquillian.container.resin.embedded_4;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
+import javax.annotation.Resource;
 
-/**
- * TestServlet
- *
- * @author <a href="mailto:aslak@redhat.com">Aslak Knutsen</a>
- * @version $Revision: $
- */
-public class TestServlet extends HttpServlet
+public class TestBean
 {
-   private static final long serialVersionUID = 1L;
+   @Resource(name = "name")
+   private String name;
 
-   public static final String URL_PATTERN = "/Test";
-
-   public static final String MESSAGE = "hello";
-
-   @Override
-   protected void service(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException
+   public String getName()
    {
-      response.getWriter().append(MESSAGE);
+      return name;
    }
 }
